centralize auth models
